Treats embedded Elixir like a comment
In the process of converting LiveView components and views into
Surface, developers can run into somewhat confusing parsing errors
if they fail to convert their Embedded Elixir `<%= foo = :bar %>`
into the Surface Elixir interpolation syntax `{{ foo = :bar }}`.

This commit adds a parsing rule to ignore embedded Elixir syntax
so that it prints to the screen, alerting the developer of the need
to change the syntax.