Support Descriptive EEx Errors
Developers often convert LiveView Components and Views into Surface
Components and Views. In the process of doing so, it is easy to forget
to change embedded elixir code to use the Surface syntax. This commit
introduces an error message to alert developers to do so.

Without this message, developers will see less specific parse errors
that do not point to the underlying cause of the error.